requirements.txt
tabulate
numpy
seaborn
matplotlib
python-docx
pymongo
flask
flask-wtf
wtforms
flask_pymongo
